Add Fluffici authentication SDK
The commit introduces the Fluffici authentication SDK which allows user authorization via OAuth. A helper class Authorize SDK was also created to handle requesting authentication and constructing URLs with specific query parameters. An AuthSDK class, to retrieve and manage client details, and a TemporaryCredentials class, to hold the authentication result and grant type, have also been added. Lastly, the required dependencies in the composer.json file and a brief README were added to explain the SDK usage and its importance.