Implement PSR-4 autoloading and expand testing suite
The changes introduce PSR-4 autoloading for main library and development files in composer.json. The core SDK files have been moved to the new src/SDK directory to better suit the new autoloading scheme. In addition, several test cases for the AuthSDK, Authorize, and TemporaryCredentials classes have been added to ensure the proper functioning and behavior of the SDK. The 'phpunit/phpunit' package was also added to the development dependencies to facilitate the running of these tests.